setup KvStore keys for limit order data structures

+++ x/dex/client/cli/query_limit_order_pool_reserve_map.go
@@ -62,7 +62,7 @@ func CmdShowLimitOrderPoolReserveMap() *cobra.Command {
 			}
 
 			params := &types.QueryGetLimitOrderPoolReserveMapRequest{
-				Count: argCount,
+				Count: uint64(argCount),
 			}
 
 			res, err := queryClient.LimitOrderPoolReserveMap(context.Background(), params)
